requirements.txt
jupyter
scipy
matplotlib
numpy
scipy
pyyaml